changes on  ANC to enable delete and update chronic illness

+++ Solutions/IQCare.Core/IQCare.Core.DbMigration/Scripts/2.2.0/33_PatientChronicillnessView.sql
@@ -1,4 +1,4 @@
-﻿IF OBJECT_ID('dbo.PatientChronicIllnessView', 'V') IS NOT NULL
+IF OBJECT_ID('dbo.PatientChronicIllnessView', 'V') IS NOT NULL
     DROP VIEW [dbo].[PatientChronicIllnessView]
 GO
 
@@ -7,4 +7,4 @@ AS
 SELECT pc.Id,pc.PatientId,pc.PatientMasterVisitId,pc.ChronicIllness AS ChronicIllnessId,lk.DisplayName AS ChronicIllness ,pc.Treatment ,pc.Dose
       ,pc.Duration ,pc.DeleteFlag,pc.CreateBy AS CreatedBy,pc.CreateDate AS DateCreated ,pc.OnsetDate ,pc.Active
   FROM PatientChronicIllness pc INNER JOIN LookupItem lk ON lk.Id = pc.ChronicIllness
-    where pc.DeleteFlag <> 1
+  where pc.DeleteFlag <> 1
